backend: (riscv) target-independent LIFO register queue
